mgr/smb: add support for new ceph vfs module
Add support for using the new ceph vfs module in Samba (aka ceph_new) to
the mgr module, and even default to using it for the samba-vfs provider.
Add two new "sub-providers" one for forcing the old vfs module
"samba-vfs/classic" and one, "samba-vfs/new", for forcing the new module
regardless of what the default for "samba-vfs" is.

Signed-off-by: John Mulligan <jmulligan@redhat.com>