mgr/cephadm: update get_container_image function
Update get_container_image to be a public function as it is only called
by code outside of its class. Add a 2nd boolean argument that forces the
use of a ceph image regardless of what the passed service name is.
This new flag will be used to create ceph image based sidecars for other
services.

Signed-off-by: John Mulligan <jmulligan@redhat.com>