mgr/cephadm: use transparent binds for haproxy not ip_nonlocal_bind
Ceph's use of haproxy and keepalived expect that the haproxy servers can
bind to IP addresses that are not already present on the system. There
are a handful of ways to do this on linux including the global sysctl
`net.ipv4.ip_nonlocal_bind=1` setting, using the socket option
IP_FREEBIND, or a few other methods (read more:
https://oswalt.dev/2022/02/non-local-address-binds-in-linux/ ).  HAproxy
supports a bind keyword `transparent` that uses one of the socket
options prior to bind allowing it to bind to an address not currently
configured on an interface when HAproxy is compiled with
USE_LINUX_TPROXY=1.  (See:
https://www.haproxy.com/documentation/haproxy-configuration-manual/2-3r1/ )

Switch the HAProxy configuration to use the transparent bind option.
Remove setting `net.ipv4.ip_nonlocal_bind=1` from cephadm.
This ought to be equivalent from an HAProxy point-of-view but should
prevent the various side effects of the global
`net.ipv4.ip_nonlocal_bind=1` setting from impacting other daemons on a
system. In particular it was found that this setting interferes with the
normal operation of the CTDB daemon.

Signed-off-by: John Mulligan <jmulligan@redhat.com>